# crypto: rework position and merkle path verification