--- a/crypto/src/proofs/groth16.rs
+++ b/crypto/src/proofs/groth16.rs
@@ -718,11 +718,10 @@ mod tests {
             let rsk = sk_sender.spend_auth_key().randomize(&spend_auth_randomizer);
             let nk = *sk_sender.nullifier_key();
             let ak = sk_sender.spend_auth_key().into();
-            let mut sct = tct::Tree::new();
-            let wrong_anchor = sct.root();
-            sct.insert(tct::Witness::Keep, note_commitment).unwrap();
+            let sct = tct::Tree::new();
             let anchor = sct.root();
-            let state_commitment_proof = sct.witness(note_commitment).unwrap();
+            let state_commitment_proof = tct::Proof::dummy(&mut OsRng, note_commitment);
+            let dummy_anchor = state_commitment_proof.root();
             // Using a random blinding factor here, but the proof will verify
             // since for dummies we only check if the value is zero, and choose
             // not to enforce the other equality constraint.
@@ -739,7 +738,7 @@ mod tests {
                 spend_auth_randomizer,
                 ak,
                 nk,
-                wrong_anchor,
+                dummy_anchor,
                 balance_commitment,
                 nf,
                 rk,