--- a/crypto/src/proofs/groth16.rs
+++ b/crypto/src/proofs/groth16.rs
@@ -774,7 +774,7 @@ mod tests {
                 Ok(self.state_commitment_proof.position())
             })?;
             merkle_path_var.verify(
-                cs.clone(),
+                cs,
                 &Boolean::TRUE,
                 position_var.inner,
                 anchor_var,