--- a/crypto/src/proofs/groth16.rs
+++ b/crypto/src/proofs/groth16.rs
@@ -772,10 +772,11 @@ mod tests {
             let position_var = tct::r1cs::PositionVar::new_witness(cs.clone(), || {
                 Ok(self.state_commitment_proof.position())
             })?;
+            let position_bits = position_var.inner.to_bits_le()?;
             merkle_path_var.verify(
                 cs,
                 &Boolean::TRUE,
-                position_var.inner,
+                &position_bits,
                 anchor_var,
                 claimed_note_commitment.inner(),
             )?;