--- a/crypto/src/proofs/groth16/delegator_vote.rs
+++ b/crypto/src/proofs/groth16/delegator_vote.rs
@@ -1,7 +1,6 @@
 use std::str::FromStr;
 
 use ark_groth16::r1cs_to_qap::LibsnarkReduction;
-use ark_r1cs_std::uint64::UInt64;
 use ark_r1cs_std::{prelude::*, uint8::UInt8};
 use ark_serialize::{CanonicalDeserialize, CanonicalSerialize};
 use decaf377::FieldExt;
@@ -75,9 +74,10 @@ impl ConstraintSynthesizer<Fq> for DelegatorVoteCircuit {
         let position_var = tct::r1cs::PositionVar::new_witness(cs.clone(), || {
             Ok(self.state_commitment_proof.position())
         })?;
-        let position_u64: u64 = self.state_commitment_proof.position().into();
-        let position_u64_var = UInt64::new_witness(cs.clone(), || Ok(position_u64))?;
-        let position_bits = position_u64_var.to_bits_le();
+        let position_bits = tct::r1cs::PositionBitsVar::new_witness(cs.clone(), || {
+            Ok(self.state_commitment_proof.position())
+        })?
+        .to_bits_le()?;
         let merkle_path_var = tct::r1cs::MerkleAuthPathVar::new_witness(cs.clone(), || {
             Ok(self.state_commitment_proof)
         })?;