--- a/crypto/src/proofs/groth16/spend.rs
+++ b/crypto/src/proofs/groth16/spend.rs
@@ -2,7 +2,9 @@ use std::str::FromStr;
 
 use ark_r1cs_std::{
     prelude::{EqGadget, FieldVar},
+    uint64::UInt64,
     uint8::UInt8,
+    ToBitsGadget,
 };
 use ark_serialize::{CanonicalDeserialize, CanonicalSerialize};
 use decaf377::FieldExt;
@@ -75,6 +77,9 @@ impl ConstraintSynthesizer<Fq> for SpendCircuit {
         let position_var = tct::r1cs::PositionVar::new_witness(cs.clone(), || {
             Ok(self.state_commitment_proof.position())
         })?;
+        let position_u64: u64 = self.state_commitment_proof.position().into();
+        let position_u64_var = UInt64::new_witness(cs.clone(), || Ok(position_u64))?;
+        let position_bits = position_u64_var.to_bits_le();
         let merkle_path_var = tct::r1cs::MerkleAuthPathVar::new_witness(cs.clone(), || {
             Ok(self.state_commitment_proof)
         })?;
@@ -113,7 +118,7 @@ impl ConstraintSynthesizer<Fq> for SpendCircuit {
         merkle_path_var.verify(
             cs.clone(),
             &is_not_dummy,
-            position_var.inner,
+            &position_bits,
             anchor_var,
             claimed_note_commitment.inner(),
         )?;