--- a/proof-params/src/gen/delegator_vote_id.rs
+++ b/proof-params/src/gen/delegator_vote_id.rs
@@ -1,3 +1,3 @@
 
-pub const PROVING_KEY_ID: &'static str = "groth16pk1afhkjvahe3zss42d4rw6yxkp5gdq8m6pxx7zrenvu7msntqlpvasn3hrt9";
-pub const VERIFICATION_KEY_ID: &'static str = "groth16vk1dawnhu3zgcrpm0m0k2ttj6946khmua99vtnjz9qy8dsp6qyqsxqquqvx55";
+pub const PROVING_KEY_ID: &'static str = "groth16pk1ud77xxwdykxrvnlyn34cktxtcly6l8fpdr9ztdq6j0kcg6qvtdrqhj0s8k";
+pub const VERIFICATION_KEY_ID: &'static str = "groth16vk1tcq3hwq4wfh6umyln9dmjw0wpzq9xlej65gxsxhr55cmg8ht286q5tawgz";