--- a/proof-params/src/gen/delegator_vote_pk.bin
+++ b/proof-params/src/gen/delegator_vote_pk.bin
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:b861f14aa2da1d42c15d23d97fc0a0a1c5949c9188542ace64bbc8db1b28d113
-size 21596592
+oid sha256:6f166de6065753d6f8ba1d42aef045224636ba520c846b87467d6ef0f89efc52
+size 21915312