--- a/proof-params/src/gen/spend_id.rs
+++ b/proof-params/src/gen/spend_id.rs
@@ -1,3 +1,3 @@
-pub const PROVING_KEY_ID: &'static str =
-    "groth16pk1vtxptxc88jjkcklwumdd7a223m278r64dzxjptpuyy9j6hvfjg7syts3e7";
-pub const VERIFICATION_KEY_ID: &'static str = "DUMMY_VALUE";
+
+pub const PROVING_KEY_ID: &'static str = "groth16pk1kwcm75dh774fz7kyua6jwzkx7ntp9vgqjrwg7mddewczn6vwcjassj2dmd";
+pub const VERIFICATION_KEY_ID: &'static str = "groth16vk1dt5lyj63cjuzx9ejv6g8l5ahs35gagjj94vquac5zzt7a36x9v4suxa58m";