--- a/proof-params/src/gen/spend_pk.bin
+++ b/proof-params/src/gen/spend_pk.bin
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:f0987ae25f904d284eeca43ac958b9d3acb9f74bd3842bb949cb22737bc40530
-size 20870352
+oid sha256:c112b6227b54187f549834faca118bdaf6bed5b2dd2a54066316c7c69bbdac96
+size 20947152