--- a/tct/src/proof.rs
+++ b/tct/src/proof.rs
@@ -55,7 +55,8 @@ impl Proof {
     /// Generate a dummy [`Proof`] for a given commitment.
     pub fn dummy<R: Rng + ?Sized>(rng: &mut R, commitment: Commitment) -> Self {
         let dummy_position = 0u64.into();
-        let dummy_auth_path: [[Hash; 3]; 24] = [[Hash::new(Fq::rand(rng)); 3]; 24];
+        //let dummy_auth_path: [[Hash; 3]; 24] = [[Hash::new(Fq::rand(rng)); 3]; 24];
+        let dummy_auth_path: [[Hash; 3]; 24] = [[Hash::new(Fq::from(0u64)); 3]; 24];
         Self::new(commitment, dummy_position, dummy_auth_path)
     }