… proof-params/src/gen/delegator_vote_vk.param
… proof-params/src/gen/spend_vk.param
… transaction/src/plan/build.rs