Closes #2135
This also resolves the last two checkboxes in #2350