@redshiftzero 2023-05-11
Performance changes compared to `main`:
* Spend
    * Proving time: 2.19s -> 2.27s
    * Number of constraints: 33484 -> 33740
* Delegator vote
    * Proving time: 2.20s -> 2.24s
    * Number of constraints: 35497 -> 36425